refactor(grouped-character): move component from experimental
BREAKING CHANGE: GroupedCharacter has new import path -
carbon-react/lib/components/grouped-character
There is a codemod available to assist with this upgrade:
npx carbon-codemod move-experimental-components <target>
See https://github.com/Sage/carbon-codemod for more information.
