+++ cypress/features/regression/accessibility/accessibilityCommon.feature
@@ -16,6 +16,7 @@ Feature: Accessibility tests - Common list
       | Configurable-items  |
       | Detail              |
       | Dialog              |
+      | GroupedCharacter    |
       | Heading             |
       | Help                |
       | Icon                |
@@ -53,3 +54,27 @@ Feature: Accessibility tests - Common list
       | Setting row      |
       | Table-ajax       |
       | Table            |
+
+  @accessibility
+  Scenario Outline: Component <component> validations boolean story page
+    When I open "<component>" component page "validations boolean" in no iframe
+    Then "<component>" component has no accessibility violations
+    Examples:
+      | component        |
+      | GroupedCharacter |
+
+  @accessibility
+  Scenario Outline: Component <component> validations string story page
+    When I open "<component>" component page "validations string component" in no iframe
+    Then "<component>" component has no accessibility violations
+    Examples:
+      | component        |
+      | GroupedCharacter |
+
+  @accessibility
+  Scenario Outline: Component <component> validations string label story page
+    When I open "<component>" component page "validations string label" in no iframe
+    Then "<component>" component has no accessibility violations
+    Examples:
+      | component        |
+      | GroupedCharacter |
