+++ cypress/fixtures/commonComponents/groupedCharacter.json
