+++ src/components/grouped-character/generate-groups.spec.js
