+++ src/components/grouped-character/grouped-character.d.ts
