+++ src/components/grouped-character/grouped-character.spec.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { mount } from "enzyme";
 import GroupedCharacter from "./grouped-character.component";
-import Label from "../label";
+import Label from "../../__experimental__/components/label";
 
 const mountComponent = (props) => mount(<GroupedCharacter {...props} />);
