+++ src/components/grouped-character/grouped-character.stories.js
@@ -3,10 +3,10 @@ import { text, object } from "@storybook/addon-knobs";
 import { action } from "@storybook/addon-actions";
 
 import GroupedCharacter from "./grouped-character.component";
-import { getCommonTextboxProps } from "../../../components/textbox/textbox.stories";
+import { getCommonTextboxProps } from "../textbox/textbox.stories";
 
 export default {
-  title: "Experimental/GroupedCharacter/Test",
+  title: "GroupedCharacter/Test",
   component: GroupedCharacter,
   parameters: {
     info: {
