+++ src/components/grouped-character/grouped-character.stories.mdx
@@ -2,14 +2,11 @@ import { useState } from "react";
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 
 import GroupedCharacter from ".";
-import { OriginalTextbox } from "../../../components/textbox";
-import Button from "../../../components/button";
-import Box from "../../../components/box";
+import { OriginalTextbox } from "../textbox";
+import Button from "../button";
+import Box from "../box";
 
-<Meta
-  title="Experimental/GroupedCharacter"
-  parameters={{ info: { disable: true } }}
-/>
+<Meta title="GroupedCharacter" parameters={{ info: { disable: true } }} />
 
 # GroupedCharacter
 
@@ -18,6 +15,7 @@ Capture data with punctuation inside the field.
 Fields can have punctuation within them (most commonly characters like - or /). This gives the user a clue about the format accepted.
 
 ## Contents
+
 - [Quick Start](#quick-start)
 - [Examples](#examples)
 - [Props](#props)
@@ -25,8 +23,9 @@ Fields can have punctuation within them (most commonly characters like - or /). 
 ## Quick Start
 
 ```javascript
-import GroupedCharacter from "carbon-react/lib/__experimental__/components/grouped-character";
+import GroupedCharacter from "carbon-react/lib/components/grouped-character";
 ```
+
 ## Examples
 
 ### Default
@@ -75,7 +74,7 @@ import GroupedCharacter from "carbon-react/lib/__experimental__/components/group
 ### AutoFocus
 
 <Preview>
-  <Story name="autoFocus" parameters={{ chromatic: { disable: true }}}>
+  <Story name="autoFocus" parameters={{ chromatic: { disable: true } }}>
     <GroupedCharacter
       label="GroupedCharacter"
       value="1231231"
@@ -90,8 +89,8 @@ import GroupedCharacter from "carbon-react/lib/__experimental__/components/group
 
 <Preview>
   <Story name="labelAlign">
-     {() => 
-      ['right', 'left'].map((alignment) => (
+    {() =>
+      ["right", "left"].map((alignment) => (
         <GroupedCharacter
           label="GroupedCharacter"
           value="1231231"
@@ -99,7 +98,7 @@ import GroupedCharacter from "carbon-react/lib/__experimental__/components/group
           separator="-"
           labelInline
           inputWidth={50}
-          key={alignment} 
+          key={alignment}
           labelAlign={alignment}
         />
       ))
@@ -132,6 +131,7 @@ import GroupedCharacter from "carbon-react/lib/__experimental__/components/group
           value="1231231"
           groups={[2, 2, 3]}
           separator={separator}
+          key={separator}
         />
       ))
     }
@@ -154,6 +154,7 @@ import GroupedCharacter from "carbon-react/lib/__experimental__/components/group
           value="1231231"
           groups={groups}
           separator="-"
+          key={groups.join("-")}
         />
       ))
     }
@@ -163,7 +164,7 @@ import GroupedCharacter from "carbon-react/lib/__experimental__/components/group
 ### With labelInline
 
 <Preview>
-  <Story name="with labelInline" parameters={{ chromatic: { disable: true }}}>
+  <Story name="with labelInline" parameters={{ chromatic: { disable: true } }}>
     <GroupedCharacter
       label="GroupedCharacter"
       value="1231231"
