+++ src/components/grouped-character/grouped-character.utils.js
