+++ src/components/grouped-character/index.d.ts
